biod: Fix memory corruption in biod_client_tool
In the call "biometrics_manager->path().value()", "path()" returns a
dbus::ObjectPath object and "value()" returns a reference to the member
string in that object. Since "path()" is an object, it is destroyed when
it goes out of scope, at which point the reference returned by "value()"
is no longer valid. Since "path()" is an rvalue (temporary) in this
case, it goes out scope after evaluation of
"biometrics_manager->path().value()".

The problem with using base::StringPiece (std::string_view) in this
situation is that it references the underlying memory, which is no
longer valid since the "path()" temporary object was destroyed. The
solution is to save a copy of the return value by replacing
base::StringPiece with std::string.

BUG=b:227148411
TEST=dut> biod_client_tool list sss
TEST=dut> valgrind biod_client_tool list sss

Change-Id: Ie6a1a3efd6aa96597fcee26bc9bb9c48fc3b5239
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3581412
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Andrea Grandi <agrandi@google.com>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>