shill: Merge DHCPv4Config into DHCPConfig (part 1)
This series of patches merges DHCPv4Config class into DHCPConfig class,
since the former one is the only inherited class of the latter one and
and thus it's confused and not necessary to keep this inheritance.

As the first step, this patch changes dhcp_config_test.cc to use
DHCPv4Config instance instead of the TestDHCPConfig, and also moves
tests in dhcpv4_config_test.cc that verifies the parameters to run the
dhcpcd into that file.

This patch only modifies unit test.

BUG=b:227560694
TEST=unit_tests

Change-Id: Id54beea3729cf1334f9ee2e7b10ce80e1083f044
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3578098
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>