cros-disks: Tighten permissions for exFAT and NTFS filesystems
exFAT and NTFS filesystems are mounted by the fuse-exfat and 3g-ntfs
fuse modules respectively. These were being used without the dmask or
fmask options (similar to umask), which caused all directories and files
on the filesystem to be permission 0777 chronos:chronos-access, which is
more permissive that required.

Pass the dmask and fmask options to these FUSE modules so that
directories and files within their filesystems are set to 0750
chronos:chronos-access.

BUG=chromium:1311173
TEST=cros_run_unit_tests --packages cros-disks
TEST=Manually ensure Files app, Crostini and ARC can read/write

Change-Id: I94c9886b4d1c16c2ca4b1e0b9b6c878911b2ae0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3578100
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>
Tested-by: Josh Simmons <simmonsjosh@google.com>