typecd: Update CableLimitingUSBSpeed to account for AMA VDOs
Currently, the CableLimitingUSBSpeed function in the Port class assumes
that a higher number in a VDO's "USB Highest Speed" field represents a
faster supported USB speed, but that is not true for AMA VDOs. This CL
updates the function to account for billboard only devices.

BUG=b:228767786
TEST=Passing all unit tests including a new test with an AMA partner
(WIMAXIT display)

Change-Id: If1f6980d4a4bcd7aee328a6cadf556618f784a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3579100
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>