cros-disks: Change MountManager::Mount()
Changed MountManager::Mount() to take a callback instead of an
input-output mount_path parameter. This is to prepare for asynchronous
notification of the mount completion (not done yet).

Since we are not passing an input mount_path anymore, some of the tests
could be simplified or even removed.

BUG=chromium:1238564
TEST=cros_workon_make cros-disks --test

Change-Id: Iec6635dbe196c2be0520f46b5824530a1a312621
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3586602
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>