shill: DHCPConfig no longer inherits IPConfig (part 1)
This series of patches will make DHCPConfig an independent class of
IPConfig. As the first step, this patch adds a new |dhcp_controller_|
field in Device, which will point to the same object which |ipconfig_|
points to, and call DHCP-related methods on this new field. As a result,
those methods can be removed from the IPConfig class.

Note that |dhcp_controller| is used as the name of DHCPConfig in the
new code since we will finally rename DHCPConfig to DHCPController.

This patch should not affect any logic.

BUG=b:227560694
TEST=unit tests

Change-Id: Iefcdfed0d2dc11acee125d8672d5157c98040cfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577667
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>