shill: Use PostTask() for DHCP callbacks
This patch modifies DHCPConfig to use PostTask() to invoke the callbacks
to Device, so that we don't need to keep a pointer to the object
itself before destroying the object (also see the comments in the header
file). This is required to make DHCPConfig no longer ref-counted.

Note that this patch may introduce some real changes since callbacks are
invoked later. This should be expected since the execution in DHCPConfig
after posting a callback should not rely on what the callback does.

BUG=b:227560694
TEST=unit_tests

Change-Id: I230d547d63dbd1cd7c9b35da9e56e9b83d999195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577669
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>