shill: DHCPProvider holds WeakPtrs to DHCPConfig
DHCPProvider need the pointers to DHCPConfig to forward the events
received from dhcpcd. Since the time that DHCPProvider need this pointer
is usually a subrange of that of Device, Device should own DHCPConfig
and DHCPProvider should only hold a WeakPtr.

Since currently the register and deregister only happen in DHCPConfig,
and DHCPConfig will always deregister itself in the destructor, it
should be safe to use a raw pointer now, and this patch will not affect
the lifetime of DHCPConfig.

BUG=b:227560694
TEST=unit_tests

Change-Id: I4e16c81638c53b741fdec81c524b8ff91223a78b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3577670
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>