diagnostics: Move executor mojom and alias namespace
Move executor.mojom from public mojom directory to
cros_healthd/executor/mojom to make it clear that it is related to
executor.

Change the namespace to chromeos.cros_healthd.mojom.

Rename ProcessResult to ExecutedProcessResult to make it clear that it
is the result for the process executed by Executor. And avoid
conflicting with ProcessResult under probe interface.

Also rename the mojo_ipc / executor_ipc to mojom in the modified files.

BUG=b:229051479
TEST=FEATURES=test emerge-zork diagnostics

Change-Id: I0a38e07e9c9134c52b7e29d2a97997ea9de1e0a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3585171
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>