shill: Remove --prepend_dns_servers flag
This patch removes the --prepend_dns_servers flag and the related logic
in shill. This flag was used to set a list of DNS servers when starting
shill in the command line, and this list will be prepended to the
IPConfig for a connection if the corresponding service is not configured
with static DNS servers. This flag was introduced for Jetstream and
should be no longer in use now.

BUG=b:229174852
TEST=unit_tests

Change-Id: I79b5ff2287226830aad89f62bc0de4245d811de3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3586608
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>