shill: cellular: apply ipv6 config from bearer statically onto netdev
Before this patch, on a dual stack cellular network, static IPv6
configuration properties from bearer is ignored and IPv6 config are
only populated by SLAAC. This interaction has a flaw that on L850
modem where bearer contains the link local address assigned by the
3GPP network, that link local address is not configured onto netdev
interface correctly. This patch fixes this issue.

Note that the interaction in this patch is based on the assumption
that SLAAC is always present - when receiving IPv6 properties from
bearer, IPv6 address is immediately configured onto netdev, but not
DNS settings, and dbus ip_config object is not updated. When SLAAC
acquires an address, that is when the SLAAC address, combining with
the DNS settings from bearer, forms the ip_config object. This
interaction is guaranteed to work on cellular, but not on IPv6 VPNs
where SLAAC might not be available. This is a temporary solution
before proper static IPv6 configuration handling in b/227563210.

BUG=b:228574659
TEST=unit
TEST=on dood(L850), verify link local address from bearer configured
TEST=on pompom(QC), verify IPv6 connectivity working as before patch

Change-Id: Ib2492e084db1df16b5277e29e577a83803bcdf4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3587950
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>