cryptohome: Move logical volume manager object into platform
Instead of creating several LVM objects/passing them through layers,
add the lvm object to platform. Move all users of lvm to the platform
object.

This also paves the way to remove the halfway fakes for encrypted
container factory and backing device factory. Once we have a proper
fake for lvm, we will get rid of those as well.

This is required to make it easier for the vault factory layers to
cache the pv/vg/lvs without needing to pass the lvm object through.

BUG=b:3588918
TEST=CQ

Change-Id: I296ccc41b5b59ff64136d5428a6b41fe339fdf3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3590356
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>