libbrillo: Add storage balloon construct
Add a construct for storage balloons. Based on the given filesystem
size at the point of inflation, the storage balloon will occupy the
remaining space to leave only a configured percentage of space available
for the remaining users of the filesystem. The current setup allows
storage balloons to be iteratively inflated: this allows us to tighten
the remaining available space as we hit the boundary of zero free disk
space.

BUG=b:234763681
TEST=unittests

Disallow-Recycled-Builds: test-failures
Change-Id: I893ddcb1fc10dd876cecb3d8944db08040c69c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3595041
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>