secanomalyd: Check logged-in state.
Only ignore '/run/arc/shared_mounts/data' W+X mounts when on the login
screen. If the mount is seen during a logged-in session, report it.

Turn secanomalyd into a DBusDaemon so that it can make D-Bus calls.
Create a SessionManagerProxy object in the Daemon class and pass it to
the SystemContext object so that it can call RetrieveActiveSessions()
on session_manager. No active sessions means the system is on the login
screen.

BUG=b:228638237
TEST=Unit tests still pass.
TEST=Deploy to device. Create /r/a/s/data mount. Check that it only
TEST=gets reported when logged_in_ == false.

Change-Id: I54c4f303e0b3a7d2421283f09b8a95aed17b72a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3601069
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Aashay Shringarpure <aashay@google.com>