arc-setup: Replace .../Downloads with .../<hash>/MyFiles/Downloads
This CL replaces a reference to .../Downloads in arc_setup.cc with
.../<hash>/MyFiles/Downloads to prepare for the deprecation of the
bind-mount of .../Downloads.

We are bypassing the presubmit check as it blocks from the missing
'oom score' fields in arc-sdcard.conf and arc-sdcard-mount.conf,
which is irrelevant to the content of this CL.
This problem is addressed in another CL: crrev.com/c/3600975.

BUG=b:174727521
TEST=tast run <DUT> arc.ContainerMount
TEST=tast run <DUT> arc.Downloads

Change-Id: I72bb746c3156e710cd47b27c58593930341f1bc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3593620
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>