hiberman: Support kernel-based encryption
This change supports systems that have been augmented with kernel-based
hibernate image encryption. It uses a new ioctl to ask the kernel to
enable encryption when going down for suspend, and gets the encrypted key
blob the kernel created in conjunction with the TPM. It uses a second
ioctl to mix in its own key contents, which will be applied only to the
hibernate data portion (and not the header or list of pagemap pages).
This way, the hibernate image is protected by both the TPM and user
authentication data.

Upon resume, the hiberman uses the same new ioctl to enable encryption
again, and hand the kernel back its encrypted key blob. The kernel
communicates with the TPM to unseal that keyblob and gets the encryption
key. Hiberman can then begin pushing the header and pagemap pages into
the kernel immediately. However it cannot push the data in, because the
user key that was mixed in is not yet available.

When the user authenticates, credentials become available to get the
user key. We use the same second new ioctl to hand that to the kernel,
and then push the encrypted hibernate data in to the kernel. The kernel
mixes its kernel key with our user key to derive the data key, and
decrypts the data we've pushed in.

A new -k commandline flag to hiberman forces hiberman not to use
kernel-based encryption (which it otherwise detects by default and uses
if available). Once the kernel encryption support has landed in the
Chromium tree, and some time goes by to ensure it's stable,
we can remove the flag and support for user-mode encryption.

BUG=b:206143128
TEST=hibernate

Change-Id: I58934e12a21c19c1bbce12969b06baf75f6d60d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3603157
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>