hiberman: Make ImageSplitter byte aware
In a world where the kernel is encrypting the hibernate image, it
interleaves chunks of pages with smaller authentication tags. In this
scheme, the final metadata and data sizes may not be even multiples of
pages.

Convert the ImageSplitter and ImageJoiner classes from working in pages
to working in bytes. Part of doing this involves simply changing the
units being tracked, and changing up the math a little bit.

Another piece that fell out is that the splitter and joiner now have to
reconcile the fact that the header may not be page aligned, but the
underlying DiskFile object still requires aligned I/O sizes. So when
reading or writing the last chunk of the header, a bounce buffer may be
temporarily needed to read/write a page aligned amount at the end of the
header file.

Additionally, the get_header_hash() used to return the number of bytes
it actually hashed, since it parsed the pagemap size out of the data
itself, and we needed to verify that wasn't tampered with. But now that
we store the image_meta_size in the metadata, and that length is handed
to ImageJoiner directly, we no longer need to compare sizes, because we
directly controlled the hash amount.

Finally, add a flag to the new() methods that can be used to disable
hashing entirely. This is currently always set to true (we want hashing
now). When the kernel is handling image integrity on its own, we won't
need to do this.

BUG=b:206143128
TEST=cargo build

Change-Id: I45df4301ea4d75f480ee5f8af6cd22da90d7d4f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3603152
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>