hiberman: Factor out header pages save/restore
Factor out some of the code in read_image() on the resume side and
write_image() on the suspend side.

On the suspend side, factor out and change the way the image move is
done in preparation for kernel-based encryption. In the
kernel-encrypted world, the image meta size (size of header and pagemap
pages) will be known ahead of time. But in a userspace encryption world,
that size is not known until the first page is moved, since it's parsed
out of the data itself. This combines with another bit of background,
which is that we don't want the ImageMover to span the divide between
header region and data region, since doing so will cause all the writes
to the data file to be misaligned if the header size ended with a
partial page.

So we split the move into three: the first (optional) move to push a
single page if the image meta size is not known, then the move of all
the header data, then the move of all the image data. We continue to
keep the header and data regions separate even in a kernel-encrypted
world so that the header region can be pushed to the kernel immediately
on resume (saving us memory and forcing the kernel to do its giant
allocation before we soak up the rest of memory preloading).

On the resume side, we mostly just factor out the header move into its
helper function. We also make the bit of code that sends in the first
byte of the data portion (to force the kernel to do its giant
allocation) unconditional, rather than conditional upon the preloader
being enabled.

Finally, fix a small bug on the resume side, which is that the code
loading the first partial page (the remaining 4095 bytes after the first
data byte was pushed in) was using an unaligned vector. DiskFiles won't
read into unaligned buffers, so this was causing failures to read the
resume image.

BUG=b:206143128
TEST=cargo build

Change-Id: Ie33bd6287591fe7752763abff5353519e158cfdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3603154
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>