hiberman: Allow ImageMover to pad last I/O
Add some logic (not yet activated) that will allow the image mover to
pad the last read or last write it does. In a kernel-encrypted world the
final image size may not be page-aligned, but the DiskFile the
ImageMover is wired to still needs page-aligned sizes. These padding
options allow the ImageMover to align the sizes for either the source or
the destination (or both), and still only use the smaller correct size
on the other end (ie don't fully move too much data).

BUG=b:206143128
TEST=cargo build

Change-Id: I6b7a7b5a72a4acf0e4554f806b040c0976f4e4ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3603156
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>