missive: Relax key delivery expectations.
Here is the story:
When Storage is initialized with no encryption key available, it will
make a key delivery request when Write call is performed. This is normal
and expected; key delivery roundtrip returns success, and then Write
proceeds to encrypt and write data to the Storage.

However, there is also a repeating upload call for most of the Storage
Queues; when such an upload is attempted, it executes if there are
recorded data (none in the case above) OR WHEN THERE IS NO KEY yet.
Such a request in most cases is attempted after Write operation is done,
but there is a possibility that it will start earlier. If that happens,
we will have two or even more key delivery trips in progress - one from
Write and one or more from repeating uploads. The way the test is
organized, only the first one will succeed, while the others will be
rejected returning error "KEY_DELIVERY UNAVAILABLE:Test uploader not
provided by the test, reason=KEY_DELIVERY" And if the one belonging to
Write gets this error, whole Write is considered failing - exactly what
we observe in the bug.

This is test-only issue (in prod multiple key delivery roundtrips will
all succeed; key updates can happen multiple times with no problem).
To fix it, I am relaxing the expectation, allowing more than one key
delivery to happen.

BUG=b:229770348
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: Ib07fdb5d2de65e5c7fe2d571a761654ccd63bbf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3601151
Reviewed-by: Zach Trudo <zatrudo@google.com>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>