cryptohome: Fetch USS experiment config on start
To launch USS experiment, we need to be able to fetch the experiment
config from a place, and that place is decided to be gstatic. The config
tells us whether this USS implementation version has known issues and
should be treated as invalid, and also how much of the total population
of the target channel should enable this experiment.

Add a UssExperimentConfigFetcher that connects to shill manager's signal
to monitor when is the device connected to the network. When connected,
get the config file from gstatic and parse its fields. In this CL we
simply logs the fields (last_invalid and population) it parsed. Metrics
reporting will be enabled in a following CL, and actual flag setting
will be added when we're ready to launch the experiment.

Add unittest for UssExperimentConfigFetcher.

BUG=b:230069013
TEST=deploy to DUT, restart and verify the logs about the config file is
correct.
TEST=verify the error logs is correct when pointing gstatic url to an
unrelated json file.
TEST=cryptohome unittests (included the added one)

Disallow-Recycled-Builds: test-failures
Cq-Depend: chromium:3607414
Change-Id: Ifa9a136e1bd9913abd17818886601c5c32041563
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3602252
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>