arc-setup: compactify Athlon CPU name parsing logic
This was suggested in https://crrev.com/c/3597565 and makes the code a
tad cleaner.

Secondarily, remove an extra line break before ") {". Because appending
to the list requires adding a "||" on the existing line anyway, this
line break trick was not making the diffs any cleaner.

BUG=b:175610620
TEST=ArcPropertyUtilTest

Change-Id: If6bcf57080b88b6292a3da7d06efd40affe53f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3602758
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Tested-by: Matthew DeVore <matvore@chromium.org>
Commit-Queue: Matthew DeVore <matvore@chromium.org>