shill: Add a comment explaining portal restart logic
The logic in Manager::SortServicesTask that triggers RestartPortalDetection
is not entirely obvious and there was some question whether the behavior
is intentional.

Add a comment explaining the logic and referencing the tracking bug until
we have test coverage and a UX decision on the expected behavior.

BUG=b:230030693
TEST=None (comment only)

Change-Id: I33e2c18971750a3ac2c4f0df34ba57b11fa8fa53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3610966
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Auto-Submit: Steven Bennetts <stevenjb@chromium.org>