tpm_manager: treat a disabled TPM as disallowed
When a TPM is active but disabled, some operations like reading nvram
spaces are still possible. But since the TPM is disabled, we can't
assume that the data is valid in that case; it might come from an old
install.

To prevent any operations from using the TPM when the TPM is disabled,
set `tpm_allowed_ = false` as if the TPM was not in the allow list. This
is only done if `USE_TPM_DYNAMIC` is enabled, so it should only affect
the reven board.

BUG=b:230018058
TEST=cros_workon_make --install tpm_manager && cros deploy dut tpm_manager
TEST=With the TPM disabled, `tpm_manager_client list_spaces` returns NVRAM_RESULT_OPERATION_DISABLED
TEST=Enrollment is no longer lost when booting Flex if the TPM is active but disabled

Change-Id: Icf375b4425f4143cf8e2df8c06a3040d08265fe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3612303
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>