vm_tools: concierge: Resolve image path names in dev mode
This is a preparation for relanding b/225278685#comment16, and does
not change non-dev-mode behavior at all. This is purely for ARCVM
developers.

When in dev mode, the Android images might be on the stateful partition
and /opt/google/vms/android/system.raw.img might be a symlink to the
stateful partition image file, /usr/local/vms/android/system.raw.img.
In that case, we need to use the resolved path so that brillo::SafeFD
calls can handle the path without errors.

The same is true for vendor.raw.image too. On the other hand, when in
production mode, we should NEVER do the special handling. In
production, the images files in /opt should NEVER ever be a symlink.
For that reason, this CL is no-op for production and does NOT resolve
any path names.

BUG=b:230760247
BUG=b:225278685
TEST=arc.Boot.vm
TEST=Run `mkdir -p /usr/local/vms/android/; cd /opt/google/vms/android;
 cp system.raw.img vendor.raw.img /usr/local/vms/android/;
 ln -sf /usr/local/vms/android/*.img .`, then run arc.Boot.vm again.

Change-Id: I1733217e4f29fe78f011ab1b208a472fc6237b9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3615609
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>