missive: Do not start periodic upload in MANUAL.
This is a true bug, however it affected only tests: manual queue was
associated with periodic upload of MAX period, which never executed
UNLESS we waited for all scheduled callbacks to execute! In this case it
was invoked and caused all kinds of troubles.

BUG=b:230741160
BUG=chromium:1312731
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: Ide57783cecd1ddcf3d93d412f39d989b8acc15f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3615614
Reviewed-by: Zach Trudo <zatrudo@google.com>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>