modemfwd: Create framework to put modem helpers into a sandbox
New devlink flash tools require cap_net_admin. Security team requires
that we only provide this capability to the necessary helpers, and that
we apply seccomp filters to limit its use.

This patch adds new modem helper attributes to indicate if the helper
requires cap_net_admin, and also allows us to provide a helper-specific
seccomp filter. These are used to configure a minijail in which the
helper is run.

CL has been tested with `cat /proc/self/status` within sandbox to ensure
capabilities and confirmed to prevent syscalls that are not included in
optional seccomp argument. I've also confirmed that fm350 flash succeeds
with seccomp filter in place, and that l850 modemfwd operation is not
hindered by the new minijail encapsulation.

BUG=b:230389055
TEST=Verified capabilities within sandbox, fw flash on brya/hatch

Change-Id: I07131d1a272826640fb63718447cb59eaa807ca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3615616
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>