power: Handle KBL toggle edge-case
When KBL goes from toggled-off to toggled-on and brightness has been
set to 0 by the user, perform a user-initiated brightness increase to
toggle back on. This avoids a situation where KBL is toggled on but
the brightness remains at 0, which a user could mistake for a
malfunctioning KBL or toggle key.

Add simple unit test for KBL toggling, that tests the typical use case.

BUG=b:212618906
TEST=Toggle KBL off/on with brightness set to several values, including
0 percent, unit test

Change-Id: I80c0c7736aae29c5ba09e35480769e3d9c0d1a11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3617035
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Roger Tinkoff <rtinkoff@google.com>
Commit-Queue: Roger Tinkoff <rtinkoff@google.com>