ippusb_bridge: Add udev change rule
An IPP-USB device will be matched by multiple sets of USB and printer
udev rules.  For the add action, the 99-ippusb rule wins and sets the
group to ippusb.  If another action subsequently arrives (e.g., change
or bind), one of the earlier rules will instead match and reset the
group.  This prevents ippusb_bridge from starting.

To preserve the group ownership, copy most of the add rule to a new rule
that triggers for everything except remove.  The RUN key is removed
because we don't want to start additional copies of ippusb_bridge for
other events.

BUG=b:230903075
TEST=udevadm trigger -s usb -c change --name-match=bus/usb/001/026

Disallow-Recycled-Builds: scarlet-cq
Change-Id: Icfcf1f36432130b9575b6e279c3514f5e515b618
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3617283
Reviewed-by: Mason Wilde <masonwilde@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>