vm_tools: Mount /usr/share/fonts in Termina-like VMs
It's reasonable that ChromeOS system fonts should be accessible in VMs
too. The way that ArcVM does this is different to the way Crostini
does.

To make this more consistent, and available to future VMs like
Bruschetta and Borealis, we instead share the fonts via concierge in the
VM setup. This will affect all termina-like VMs but they will need to
mount the directory (as a virtio fs) themselves.

BUG=b:214865417
TEST=emerge and run `mount -t virtiofs fonts /some/dir` in the vm

Change-Id: I30721e9e5fc4d77e6a976d26a6e38540534fd752
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3616549
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Tested-by: Nic Hollingum <hollingum@google.com>