cryptohome: Don't use SecureBlob for wrapped USS
The wrapped user secret stash (the USS container) doesn't store any
secrets in clear and is also persisted on disk as-is. Given that, and as
SecureBlob has some nonzero overhead, this commit updates the code and
the serialization layer to use simple Blob for the USS container.

This commit also fixes the warnings produced by the flatbuffer code
generator about mixing secure and insecure blobs.

Note that the USS payload, which contains secrets, continues to use
SecureBlob - this stays intact.

BUG=b:218943323
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I7e49b10b8c17ec3dab835e78b132f5d7ffefbf7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3620404
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>