power: Add Charge History to Adaptive Charging
Charge History is used for tracking when, and for how long, chargers
are connected to the system. For privacy reasons, since this
information is stored in /var/lib/, We obfuscate the timestamps by
flooring everything to the nearest 15 minute multiple. We also retain
data for 30 days and at most 50 charger plug/unplug events.

BUG=chromium:1243163
TEST=manual and unittests

Change-Id: I2266724eed7b74e38481133168a37adc3438b0dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3607181
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>