power: Add Heuristic for enabling Adaptive Charging
This adds a heuristic which makes Adaptive Charging run only when there
is sufficient charge history for the system, and when most of the time
spent on the AC charger is with a full charge. Whether this heuristic is
allowing Adaptive Charging is reported via the PowerSupplyProperties
proto via DBus.

BUG=chromium:1243163
TEST=manual and unittests

Change-Id: Iaf79fd60ef0ec93c87c0c1702249bf569da71bf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3631671
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>