libec: Split MkbpWakeMaskCommand into a Get and Set class
When getting the wake mask, there is a response from the EC, but when
setting the wake mask there is no response. Since libec validates the
response size when running the command, the previous implementation was
returning a failure since it expected a response size that matched the
size of "struct ec_response_mkbp_event_wake_mask".

BUG=b:231619554, b:144959033
TEST=FEATURES="test" emerge-hatch libec
TEST=On CoachZ:
     echo 1 > /var/lib/power_manager/wake_on_dp
     restart powerd
     ectool mkbpwakemask get event
     MBKP event wake mask: 0x00000406
TEST=On CoachZ:
     echo 0 > /var/lib/power_manager/wake_on_dp
     restart powerd
     ectool mkbpwakemask get event
     MBKP event wake mask: 0x00000006

Change-Id: I0e3e8301176775ec627025ef3bed41ac99020d29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3632816
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>