power: Reduce base::RunLoop::RunUntilIdle() usage
This CL replaces some of the base::RunLoop::RunUntilIdle() with
base::RunLoop::Run() and base::RunLoop()::QuitClosure() in
SensorServiceHandlerTest, and remove unnecessary ones in
AmbientLightSensorManagerMojoTest.

There are still some of the usages that are not simple to be replaced
with other mechanisms.

BUG=b:229668577
TEST=unit tests

Change-Id: I2a994be19379fc66876f89733cfaa486ccdf9204
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3629893
Auto-Submit: Cheng-Hao Yang <chenghaoyang@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Cheng-Hao Yang <chenghaoyang@chromium.org>